requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
httpx==0.27.0
tqdm==4.66.5

# vector + pdf + embeddings
chromadb==0.5.5
pypdf==4.3.1
sentence-transformers==3.0.1
cryptography>=45.0.0

# LLM (Claude)
anthropic==0.37.1

# uploads (images)
python-multipart==0.0.9
pillow==10.4.0
